refactor(llm prompting): Rename `local_llm_prompting` to `llm_prompting`
